% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualListBox.R
\name{dualListBoxInput}
\alias{dualListBoxInput}
\alias{updateDualListBoxInput}
\title{React Dual Listbox}
\usage{
dualListBoxInput(inputId, label, ..., width = NULL)

updateDualListBoxInput(session, inputId, value, configuration = NULL)
}
\description{
A feature-rich dual listbox.
}
