% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactSelect.R
\name{reactSelectInput}
\alias{reactSelectInput}
\alias{prepareReactSelectOptions}
\title{React Select}
\usage{
reactSelectInput(
  inputId,
  label,
  options,
  selected = NULL,
  placeholder = "Select...",
  isMulti = FALSE,
  closeMenuOnSelect = !isMulti,
  isSearchable = FALSE,
  isDisabled = FALSE,
  width = 500
)

prepareReactSelectOptions(choices)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{options}{Specify the options the user can select from.}

\item{selected}{Control the current value.}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this
option.}

\item{isMulti}{Allow the user to select multiple values.}

\item{isSearchable}{Allow the user to search for matching options.}

\item{isDisabled}{Disable the control.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:validateCssUnit]{validateCssUnit()}}.}

\item{choices}{A character vector or a named list, similar to \code{\link[shiny:selectInput]{shiny::selectInput()}}'s choices argument.}
}
\description{
A feature-rich dual listbox.
}
