% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualListBox.R
\name{dualListBoxInput}
\alias{dualListBoxInput}
\alias{updateDualListBoxInput}
\alias{prepareDualListBoxOptions}
\title{React Dual Listbox}
\usage{
dualListBoxInput(
  inputId,
  label,
  options,
  selected = NULL,
  canFilter = FALSE,
  showOrderButtons = FALSE,
  preserveSelectOrder = FALSE,
  alignActions = c("middle", "top"),
  width = 500
)

updateDualListBoxInput(
  inputId,
  selected,
  configuration = NULL,
  session = shiny::getDefaultReactiveDomain()
)

prepareDualListBoxOptions(choices)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{options}{Specifies the list of options that may exist on either side of the dual list box.}

\item{selected}{A list of the selected options appearing in the rightmost list box.}

\item{canFilter}{If \code{TRUE}, search boxes will appear above both list boxes, allowing the user to filter the results.}

\item{showOrderButtons}{If true, a set of up/down buttons will appear near the selected list box to allow the user to re-arrange the items.}

\item{preserveSelectOrder}{If \code{TRUE}, the order in which the available options are selected are preserved when the items are moved to the right.}

\item{alignActions}{A value specifying whether to align the action buttons to the 'top' or 'middle'.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:validateCssUnit]{validateCssUnit()}}.}

\item{configuration}{Other options to update the dual list box.}

\item{session}{The Shiny session.}

\item{choices}{A character vector or a named list, similar to \code{\link[shiny:selectInput]{shiny::selectInput()}}'s choices argument.}
}
\description{
A feature-rich dual listbox, https://jakezatecky.github.io/react-dual-listbox/.
}
\examples{
library(shiny)
library(bslib)
library(reactwidgets)

ui <- page_fluid(
  titlePanel("dualListBoxInput Example"),
  fluidRow(
    column(
      width = 6,
      dualListBoxInput(
        "ID1",
        label = "Basic Dual list box example:",
        options = prepareDualListBoxOptions(c(
          "luna", "phobos", "deimos", "io", "europa", "ganymede", "callisto",
          "mimas", "enceladus", "tethys", "rhea", "titan", "iapetus"
        )),
        selected = "tethys",
        width = "100\%"
      ),
      verbatimTextOutput("res1")
    ),
    column(
      width = 6,
      dualListBoxInput(
        "ID2",
        label = "Dual list box with options example:",
        options = prepareDualListBoxOptions(list(
          earth = list("luna"),
          mars = c("phobos", "deimos"),
          jupiter = c("io", "europa", "ganymede", "callisto")
        )),
        canFilter = TRUE,
        showOrderButtons = TRUE,
        preserveSelectOrder = TRUE,
        alignActions = "top",
        width = "100\%"
      ),
      verbatimTextOutput("res2")
    )
  )
)

server <- function(input, output, session) {
  output$res1 <- renderPrint({
    input$ID1
  })
  output$res2 <- renderPrint({
    input$ID2
  })
}

if (interactive())
  shinyApp(ui, server)
}
