% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactSelect.R
\name{reactSelectInput}
\alias{reactSelectInput}
\alias{prepareReactSelectOptions}
\title{React Select}
\usage{
reactSelectInput(
  inputId,
  label,
  options,
  selected = NULL,
  placeholder = "Select...",
  isMulti = FALSE,
  closeMenuOnSelect = !isMulti,
  isSearchable = FALSE,
  isDisabled = FALSE,
  width = 500
)

prepareReactSelectOptions(choices)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{options}{Specify the options the user can select from.}

\item{selected}{Control the current value.}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this
option.}

\item{isMulti}{Allow the user to select multiple values.}

\item{closeMenuOnSelect}{Close the select menu when the user selects an option.}

\item{isSearchable}{Allow the user to search for matching options.}

\item{isDisabled}{Disable the control.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:validateCssUnit]{validateCssUnit()}}.}

\item{choices}{A character vector or a named list, similar to \code{\link[shiny:selectInput]{shiny::selectInput()}}'s choices argument.}
}
\description{
A feature-rich dual listbox.
}
\examples{
library(shiny)
library(bslib)
library(reactwidgets)

ui <- page_fluid(
  titlePanel("reactSelectInput Example"),
  fluidRow(
    column(
      width = 6,
      reactSelectInput(
        "ID1",
        label = "Basic react select example:",
        options = prepareReactSelectOptions(c(
          "luna", "phobos", "deimos", "io", "europa", "ganymede", "callisto",
          "mimas", "enceladus", "tethys", "rhea", "titan", "iapetus"
        )),
        width = "100\%"
      ),
      verbatimTextOutput("res1")
    ),
    column(
      width = 6,
      reactSelectInput(
        "ID2",
        label = "Multi react select example:",
        options = prepareReactSelectOptions(list(
          earth = list("luna"),
          mars = c("phobos", "deimos"),
          jupiter = c("io", "europa", "ganymede", "callisto")
        )),
        selected = "ganymede",
        placeholder = "Search for an option by typing",
        isMulti = TRUE,
        isSearchable = TRUE,
        width = "100\%"
      ),
      verbatimTextOutput("res2")
    )
  )
)

server <- function(input, output, session) {
  output$res1 <- renderPrint({
    input$ID1
  })
  output$res2 <- renderPrint({
    input$ID2
  })
}

if (interactive())
  shinyApp(ui, server)
}
